[Octavia] Add support for dual CAs
Best practice when deploying TLS certificates for use by Octavia is to
use two seperate CAs, in order to avoid the possibility of Amphorae
being able to impersonate a controller by using the same server CA to
sign certificates.

This commit updates the relevant Kolla-Ansible configuration and
template in order to handle the additional files / options.