Fix idempotency of fluentd customisations
If we have a custom fluentd input/output/filter config with the same name as a
default config, we will first write out the default config then replace it with
the custom one, resulting in two changed tasks. This change fixes this by
skipping default configs that have been customised.