Bump HAProxy max connection limits (#71)
* Bump HAProxy max connection limits

The default connection limits for backends appears to be 2000.

However, mariadb defaults to a max of 10000 conections, so we up
the default limit to match the mariadb limit.

We could patch kayobe config's kolla/globals.yml to update the
global max connection limit, but this seems a more sensible default.

Fixes: HC-1406

* Add missing maxcon for second default block