Adding Prometheus Blackbox Exporter.
A custom blackbox-exporter module has been added called os_endpoint
which allows for the probing of endpoints over HTTP and HTTPS.
This is required because all (at least core) OpenStack endpoints
return a status code of either 200 or 300. This module checks
for the word 'versions' and that the status code matches.

This change introduces a new variable `prometheus_blackbox_exporter_endpoints`.
Currently no defaults are specified because the configuration is heavily
dependent on the deployment.

Blackbox exporter targets are the end points that are then transformed
by relabel_configs rules into URLs to trigger a probe. You do not specify
the Blackbox exporter itself as a target, as each endpoint will be
transformed into a Blackbox URL target which Prometheus then calls.

The Blackbox exporter binds to the api interface address and not
localhost. This also makes the placement of the service more flexible.

In the example provided, there is support to specify the blackbox
module to used as part of the target configuration, it been built
into the target itself to avoid the need for multiple scrape jobs.
A label is also included to help identify the metric.